+++ lib/swagger2ts.d.ts
@@ -1,8 +1,0 @@
-import { SwaggerParser } from './cli';
-export default function swagger2ts(
-  swaggerParser: SwaggerParser,
-  clear?: boolean
-): Promise<{
-  code: Number;
-  message?: string;
-}>;