+++ src/consts.ts
@@ -1,0 +1,4 @@
+import * as path from 'path';
+
+export const pluginsPath = path.join(__dirname, '..', 'plugins');
+export const generatorPath = path.join(pluginsPath, 'swagger-codegen-cli.jar');